% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Bicluster-class}
\alias{Bicluster-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{block}}{ANY.}

\item{\code{CoReg_gene}}{ANY.}

\item{\code{CoCond_cell}}{ANY.}
}}

