% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMeta.R
\name{.addMeta}
\alias{.addMeta}
\title{Create addmeta function for adding meta information}
\usage{
.addMeta(object = NULL, meta.info = NULL)
}
\arguments{
\item{object}{BRIC object}

\item{meta.info}{meta information table should be a data frame with rows representing cell and coloumn representing different group condition}
}
\value{

}
\description{
Create addmeta function for adding meta information
}
