% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runBicluster}
\alias{.runBicluster}
\title{Run cluster}
\usage{
.runBicluster(
  object = NULL,
  DiscretizationModel = "LTMG",
  OpenDual = TRUE,
  Extension = 0.9,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)
}
\arguments{
\item{BlockCellMin}{}
}
\value{

}
\description{
Run cluster
}
