% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runBiclusterBaseOnDiscretization}
\alias{.runBiclusterBaseOnDiscretization}
\title{Title}
\usage{
.runBiclusterBaseOnDiscretization(
  object = NULL,
  OpenDual = TRUE,
  Extension = 0.9,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)
}
\description{
Title
}
