% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bicluster.R
\name{.runDiscretization}
\alias{.runDiscretization}
\title{run discretization}
\usage{
.runDiscretization(object = NULL, q = 0.06, LogTransformation = FALSE)
}
\arguments{
\item{q}{}
}
\description{
run discretization
}
