% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{LTMGr-class}
\alias{LTMGr-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{LTMG_discrete}}{matrix.}

\item{\code{LTMG_BinarySingleSignal}}{matrix.}

\item{\code{LTMG_BinaryMultisignal}}{matrix.}

\item{\code{.}}{}
}}

