% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DimensionReducntionBasedOnLTMG.R
\name{PlotDimension}
\alias{PlotDimension}
\alias{.plotDimension}
\alias{PlotDimension,BRIC-method}
\title{Title Visualize dimension reduction results}
\usage{
PlotDimension(object, ...)

.plotDimension(object, reduction = "umap", pt_size = 1)

\S4method{PlotDimension}{BRIC}(object, reduction = "umap", pt_size = 1)
}
\arguments{
\item{object}{Input Object}

\item{reduction}{Choose one of approaches for dimension reduction, including "pca", "tsne", "umap".}

\item{pt_size}{point size, default is 0.}
}
\value{

}
\description{
Title Visualize dimension reduction results
}
