% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotHeatmap.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\alias{.plotHeatmap}
\alias{PlotHeatmap,BRIC-method}
\title{plot heatmap based on bicluster}
\usage{
PlotHeatmap(object, ...)

.plotHeatmap(object = object, N.block = 2)

\S4method{PlotHeatmap}{BRIC}(object = object, N.block = 2)
}
\arguments{
\item{object}{BRIC object}

\item{N.block}{number of blocks}
}
\description{
plot heatmap based on bicluster
}
