% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AddMeta.R
\name{PlotMeta}
\alias{PlotMeta}
\alias{.plotMeta}
\alias{PlotMeta,BRIC-method}
\title{Plot violin plot based on meta data.}
\usage{
PlotMeta(object, ...)

.plotMeta(object = NULL)

\S4method{PlotMeta}{BRIC}(object = NULL)
}
\value{

}
\description{
Plot violin plot based on meta data.
}
