% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotModuleNetwork}
\alias{PlotModuleNetwork}
\alias{.plotmodulenetwork}
\alias{PlotModuleNetwork,BRIC-method}
\title{Title PlotModuleNetwork}
\usage{
PlotModuleNetwork(object, ...)

.plotmodulenetwork(
  object = NULL,
  N.bicluster = c(1, 5),
  Node.color = "#E8E504",
  node.label.cex = 1
)

\S4method{PlotModuleNetwork}{BRIC}(
  object = NULL,
  N.bicluster = c(1, 5),
  Node.color = "#E8E504",
  node.label.cex = 1
)
}
\arguments{
\item{object}{Input object.}

\item{N.bicluster}{number of biclsuter to plot.}

\item{Node.color}{color of nodes. This parameter also accepts color codes, e.g. "#AE1503" or "darkred."}

\item{BRIC}{}
}
\value{

}
\description{
This function will visualize
}
