% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\alias{.qunetwork}
\alias{PlotNetwork,BRIC-method}
\title{plot network based on bicluster}
\usage{
PlotNetwork(object, ...)

.qunetwork(
  object = NULL,
  Bic.index = 1,
  method = c("pearson", "kendall", "spearman"),
  is.plot = TRUE
)

\S4method{PlotNetwork}{BRIC}(
  object = NULL,
  Bic.index = 1,
  method = c("pearson", "kendall", "spearman"),
  is.plot = TRUE
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
plot network based on bicluster
}
