% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PreprocessData.R
\name{ProcessData}
\alias{ProcessData}
\alias{.processData}
\alias{ProcessData,BRIC-method}
\title{Process data}
\usage{
ProcessData(object, ...)

.processData(
  object = NULL,
  normalization = "LibrarySizeNormalization",
  IsImputation = FALSE,
  seed = 123
)

\S4method{ProcessData}{BRIC}(
  object = NULL,
  normalization = "LibrarySizeNormalization",
  IsImputation = FALSE,
  seed = 123
)
}
\arguments{
\item{object}{}

\item{normalization}{two options: (1)library size noralization by using library size factor: 1e6, equal to CPM, or (2) using "scran" normalization method.}

\item{IsImputation}{imputation method is provided by DrImpute. Default is FALSE.}

\item{seed}{set seed for reproducibility.}
}
\description{
Process data via normalization and imputation.
}
