% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{ReadFrom10X_h5}
\alias{ReadFrom10X_h5}
\title{Read 10X HDF5 file based on Seurat package}
\usage{
ReadFrom10X_h5(input = NULL, ...)
}
\arguments{
\item{...}{inherit function from Seurat::Read10X_h5}

\item{Input}{input 10X Chromium output data with the extented name as ".h5"}
}
\value{
The output from \code{\link{Read10X_h5}}
}
\description{
Read 10X HDF5 file based on Seurat package
}
\examples{
\dontrun{
my.matrix<-ReadFrom10X_h5("path/to/data/files")
}

}
