% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{ReadFromMat}
\alias{ReadFromMat}
\title{Read in gene expression matrix: cell in columns and gene in rows}
\usage{
ReadFromMat(input = NULL, sep = "\\t", header = TRUE, rownames = 1, ...)
}
\arguments{
\item{sep}{defauld separator is tab.}

\item{...}{inherit function from read.delim}

\item{Input}{input gene expression matrix.}
}
\value{

}
\description{
Read in gene expression matrix: cell in columns and gene in rows
}
