% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Bicluster.R
\name{RunBicluster}
\alias{RunBicluster}
\alias{.runBicluster}
\alias{RunBicluster,BRIC-method}
\title{Run cluster}
\usage{
RunBicluster(object, ...)

.runBicluster(
  object = NULL,
  DiscretizationModel = "LTMG",
  OpenDual = TRUE,
  Extension = 0.9,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)

\S4method{RunBicluster}{BRIC}(
  object = NULL,
  DiscretizationModel = "LTMG",
  OpenDual = TRUE,
  Extension = 0.9,
  NumBlockOutput = 100,
  BlockOverlap = 0.7,
  BlockCellMin = 15
)
}
\arguments{
\item{DiscretizationModel}{asdasdasdasdasd}
}
\value{

}
\description{
Run cluster
}
