% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,BRIC-method}
\title{RunLTMG}
\usage{
RunLTMG(object, Gene_use, k = 5)

.RunLTMG(object, Gene_use, k = 5)

\S4method{RunLTMG}{BRIC}(object, Gene_use, k = 5)
}
\value{

}
\description{
RunLTMG
}
