% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AddMeta.R
\name{SubsetData}
\alias{SubsetData}
\alias{.subset_data}
\alias{SubsetData,BRIC-method}
\title{Title subset data by number of count and numebr of features.}
\usage{
SubsetData(object, ...)

.subset_data(
  object,
  nFeature.upper = Inf,
  nFeature.lower = -Inf,
  Counts.upper = Inf,
  Counts.lower = -Inf
)

\S4method{SubsetData}{BRIC}(
  object,
  nFeature.upper = Inf,
  nFeature.lower = -Inf,
  Counts.upper = Inf,
  Counts.lower = -Inf
)
}
\arguments{
\item{nFeature.upper}{select upper limit for number of feature}

\item{nFeature.lower}{select lower limit for number of feature}

\item{Counts.upper}{select upper limit for number of UMI counts}

\item{Counts.lower}{select lower limit for number of UMI counts}
}
\value{

}
\description{
Title subset data by number of count and numebr of features.
}
