% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateBRICObject}
\alias{CreateBRICObject}
\title{Title}
\usage{
CreateBRICObject(
  x = input_matrix,
  LTMGr = new(Class = "LTMGr"),
  Bicluster = new(Class = "Bicluster")
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
