% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\alias{.plotnetwork}
\alias{PlotNetwork,BRIC-method}
\title{Title PlotNetwork}
\usage{
PlotNetwork(object, ...)

.plotnetwork(
  object,
  edge.by = "gene",
  lay.out = "linear",
  N.bicluster = c(1:20),
  ...
)

\S4method{PlotNetwork}{BRIC}(
  object,
  edge.by = "gene",
  lay.out = "linear",
  N.bicluster = c(1:20),
  ...
)
}
\arguments{
\item{object}{Input object}

\item{...}{}

\item{edge.by}{This parameter decide the edge by "cell" or by "gene". The default value is by gene.}

\item{lay.out}{The type of layout to create, include linear (default), circle, kk.}

\item{BRIC}{}
}
\value{

}
\description{
This function is for building the module- module network. Nodes mean individual module. Edges mean overlapped gene (or cell).
The weight of line show the number of gene (or cell).
}
