% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,BRIC-method}
\title{RunLTMG}
\usage{
RunLTMG(object, NFeatures)

.RunLTMG(object = NULL, NFeatures = NULL)

\S4method{RunLTMG}{BRIC}(object = NULL, NFeatures = NULL)
}
\value{

}
\description{
RunLTMG
}
