% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,BRIC-method}
\title{RunLTMG}
\usage{
RunLTMG(object, ...)

.RunLTMG(object, Gene_use = NULL, seed = 123, k = 5)

\S4method{RunLTMG}{BRIC}(object, Gene_use = NULL, seed = 123, k = 5)
}
\arguments{
\item{Gene_use}{using X numebr of top variant gene. input a number, recommend 2000.}
}
\value{

}
\description{
This function is for moduling gene expression data via Left-truncated mixture Guassian model.
}
